improved logic in download_history_files

+++ utils.py
@@ -93,40 +93,43 @@ def is_old(filename: str) -> bool:
         return True
     return False
 
-def run_containment(dir_path: str|Path):
+def run_containment(dir_path: str|Path) -> bool:
     
     if not isinstance(dir_path, Path):
         dir_path = Path(dir_path)
 
-    files_with_sizes = [(p, p.stat().st_size) for p in dir_path.glob("*") if p.is_file()]
-    # Sort by size (descending)
-    files_sorted = sorted(files_with_sizes, key=lambda x: x[0], reverse=True)
-    files_sorted = sorted(files_with_sizes, key=lambda x: x[1], reverse=True)
-    to_be_deleted = set()
+    keep_containing = True
+    while keep_containing:
+        files_with_sizes = [(p, p.stat().st_size) for p in dir_path.glob("*") if p.is_file()]
+        # Sort by size (descending)
+        files_sorted = sorted(files_with_sizes, key=lambda x: x[0], reverse=True)
+        files_sorted = sorted(files_with_sizes, key=lambda x: x[1], reverse=True)
+        to_be_deleted = set()
 
-    for big_file in files_sorted:
-        for small_file in reversed(files_sorted):
-            if small_file is big_file:
-                break #stop when reachin the same file
-            with open(dir_path / small_file[0].name, "rb") as f1, open(dir_path / big_file[0].name, "rb") as f2:
-                small = f1.read()
-                big = f2.read()
-                if is_equal(small, big):
-                    if is_old(big_file[0].name): 
-                        
-                        print(f"{colored(small_file[0].name, 'red', attrs=['bold'])} will be deleted (equal to {colored(big_file[0].name, 'green', attrs=['bold'])})")
+        for big_file in files_sorted:
+            for small_file in reversed(files_sorted):
+                if small_file is big_file:
+                    break #stop when reachin the same file
+                with open(dir_path / small_file[0].name, "rb") as f1, open(dir_path / big_file[0].name, "rb") as f2:
+                    small = f1.read()
+                    big = f2.read()
+                    if is_equal(small, big):
+                        if is_old(big_file[0].name): 
+                            
+                            print(f"{colored(small_file[0].name, 'red', attrs=['bold'])} will be deleted (equal to {colored(big_file[0].name, 'green', attrs=['bold'])})")
+                            to_be_deleted.add(dir_path / small_file[0].name)
+
+                        else: 
+
+                            print(f"{colored(big_file[0].name, 'red', attrs=['bold'])} will be deleted (equal to {colored(small_file[0].name, 'green', attrs=['bold'])})")
+                            to_be_deleted.add(dir_path / big_file[0].name)
+
+                    elif is_contained(small, big):
+                        print(f"{colored(small_file[0].name, 'red', attrs=['bold'])} will be deleted (contained in {colored(big_file[0].name, 'green', attrs=['bold'])})")
                         to_be_deleted.add(dir_path / small_file[0].name)
-
-                    else: 
-
-                        print(f"{colored(big_file[0].name, 'red', attrs=['bold'])} will be deleted (equal to {colored(small_file[0].name, 'green', attrs=['bold'])})")
-                        to_be_deleted.add(dir_path / big_file[0].name)
-
-                elif is_contained(small, big):
-                    print(f"{colored(small_file[0].name, 'red', attrs=['bold'])} will be deleted (contained in {colored(big_file[0].name, 'green', attrs=['bold'])})")
-                    to_be_deleted.add(dir_path / small_file[0].name)
-                    
-    for file_path in to_be_deleted:
-        file_path.unlink()
-
-    # TODO: compact backup files
+                        
+        if not to_be_deleted:
+            keep_containing = False
+        else:
+            for file_path in to_be_deleted:
+                    file_path.unlink()
